+++ dev/isolated_test.sh
@@ -13,7 +13,7 @@ trap 'set +e; popd; rm -rf "$TEMPDIR"' EXIT
 rsync -a tests "$TEMPDIR" --exclude=__pycache__
 
 pushd "$TEMPDIR"
-python3 -m venv venv
-venv/bin/python3 -m pip -q install "$DISTFILE"
-venv/bin/python3 -Im unittest
+python -m venv venv
+venv/bin/python -m pip -q install "$DISTFILE"
+venv/bin/python -Im unittest