+++ tests/test_dt.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python
 """Tests for ``igbpyutils.dt``.
 
 Author, Copyright, and License